refactor(stella-tui): split the deck envelope so the next protocol message has room
`crates/stella-tui/src/envelope.rs` sat at 1485 lines, fifteen under the
ceiling `scripts/check-file-size.sh` enforces, and the file has no baseline
entry — so the next deck protocol message would have failed the gate inside a
change about something else, with a split as the only remedy.

Split it into `envelope.rs` beside the `envelope/` folder it already had, along
the seam the file's own ordering shows:

- `envelope/inbound.rs` — driver to deck. `Inbound`, plus `SessionPhase`,
  `SessionInfo`, `NotificationInfo`, `SplashCue`, `AgentScope`,
  `AgentVersionInfo` and `InstalledAgentEntry`, each carried by one `Inbound`
  case and nothing else.
- `envelope/workspace_input.rs` — deck to driver. `WorkspaceInput`, plus
  `KeepStrength`, `Secret` and `AgentControl`.
- `envelope.rs` keeps `AgentId`, `AgentMeta`, `AgentStatus` and `PanelSeat`,
  which both halves use, and re-exports every name at the path it had.

No public path moved: `crates/stella-tui/src/lib.rs` is untouched and
`cargo check -p stella-cli --all-targets` compiles with no import edited.

## The second fix: the prose ratchet could not see a split

`scripts/check-prose.py` follows a file that MOVES, by reading git's rename
pairs. A split produces none — the source file is still there, so git reports a
modify beside an add. The sentences that moved into the new files were charged
to this branch, and a new file is held to reading grade 6.00, so the split above
was asked for a rewrite of 1300 lines of doc comments nobody had opened. That
made `mod` hierarchy, the remedy AGENTS.md names for a crowded file,
unaffordable.

`split_sources` pairs them instead: a new file whose prose lines mostly came out
of a file the same change shrank is judged against that file's numbers in the
base tree. Requiring the lines to be gone from the source is what stops a copy
claiming one allowance twice, and prose written during the split is in neither
line set — so it lowers the share and, past the threshold, carries nothing.

`scripts/test-prose-guard.sh`'s new X block covers it. X1 is the witness: its
fixture exits 1 on the guard at `main` and 0 on this one.

## Prose

Both new files came in under the source's own reading grade rather than at it,
and the pattern findings the split carried are gone rather than moved: the
`filler-adverb`, `historical-reference`, `honesty-declaration`,
`interface-jargon`, `issue-reference` and `rust-jargon` hits in the moved text
are rewritten, so `scripts/prose-baseline.txt` needed no entry and gains none.

Closes #5973
