+++ crates/stella-tui/src/envelope.rs
@@ -10,21 +10,44 @@
 //! pure fold of *its* `AgentEvent`s; the envelope only adds the routing tag the
 //! deck needs to keep N folds side by side.
 
-use std::collections::BTreeMap;
-
-use stella_protocol::AgentEvent;
-
 pub(crate) mod broadcast;
+mod engine_config;
+/// Driver → deck: the [`Inbound`] envelope and the read models only it carries.
+mod inbound;
 mod inspect;
+mod issues;
+/// The MCP tab's read models (rows, search results, inspector detail).
+mod mcp;
+/// The `/models` role table: the open role vocabulary and its render order.
+pub mod roles;
+mod skills;
+/// Why Esc steers rather than cancels — doc-only.
+pub mod steering;
+mod tool_policy;
+/// Deck → driver: the [`WorkspaceInput`] envelope and the types only it carries.
+mod workspace_input;
 
+// Every name below keeps the public path it had when this module was one
+// file, so no caller learns where anything went.
 pub use broadcast::Broadcast;
+pub use engine_config::{EngineAgentState, EngineConfigState, EngineRole, RoleWiringRow, SeatRow};
+pub use inbound::{
+    AgentScope, AgentVersionInfo, Inbound, InstalledAgentEntry, NotificationInfo, SessionInfo,
+    SessionPhase, SplashCue,
+};
 pub use inspect::{InspectMessage, InspectSection, InspectView, JournalEra, RecordedCallInfo};
-
-use stella_tool_facts::readiness::IndexReadiness;
-
-use crate::graph::GraphSnapshot;
-use crate::input::UserInput;
-use crate::start_work::StartWorkDraft;
+pub use issues::{EntityField, EntityHit, IssueAction, IssueRow, LinkedWork};
+pub use mcp::{
+    GRAPH_SERVER, McpLiveIdentity, McpLookupState, McpSearchItem, McpSearchOutcome,
+    McpServerDetail, McpServerInfo, McpSignature, McpSourceTier, McpToolRow,
+};
+pub use roles::{RoleTableEntry, role_table};
+pub use skills::{
+    LearnedProvenance, LearnedSource, RejectedSkillRow, SkillOp, SkillRow, SkillScope,
+    SkillSearchHit, SkillsView,
+};
+pub use tool_policy::{ToolDenial, ToolPolicyState, ToolRow, ToolScope};
+pub use workspace_input::{AgentControl, KeepStrength, Secret, WorkspaceInput};
 
 /// Stable identifier for one agent/run within the workspace. Human-meaningful
 /// where possible (`"lead"`, `"sub:auth-refactor"`) — it is shown on screen, so
@@ -168,1290 +191,6 @@ pub struct PanelSeat {
     /// admitted. `None` on every other surface, and on a name that collided
     /// with a built-in — a refused name is a notice, never a quiet seat.
     pub command: Option<String>,
-}
-
-/// One item on the workspace inbound channel — the multi-agent envelope the
-/// deck folds. `--output-format stream-json` remains one `AgentEvent` per line
-/// per agent; the envelope only adds the routing tag the deck needs.
-#[derive(Clone, Debug)]
-pub enum Inbound {
-    /// A new agent joined the workspace — its dashboard row appears.
-    Register(AgentMeta),
-    /// Remove one agent's dashboard row — the visual-lifecycle inverse of
-    /// [`Inbound::Register`]. Presentation only: journaling and model state
-    /// are unaffected (the removed lane's history stays in its own session's
-    /// journal, and the engine's conversation is untouched). The driver only
-    /// ever sends it for lanes with no live process behind them — e.g. the
-    /// terminal worker rows of a session the deck is navigating away from,
-    /// which must not linger on the next session's dashboard. Folding an
-    /// unknown id is a no-op.
-    Deregister { agent: AgentId },
-    /// An `AgentEvent` belonging to one agent.
-    Event { agent: AgentId, event: AgentEvent },
-    /// A local `$` shell command's output, folded into an EXISTING lane's
-    /// transcript so it reads inline in the session the user is looking at.
-    ///
-    /// Deliberately not [`Inbound::Event`]: that path derives agent status
-    /// from the event, and `ToolStart`/`ToolResult` both mean
-    /// [`AgentStatus::Running`] — so routing a `$` command through it would
-    /// flip an idle agent to Running and leave it spinning forever (nothing
-    /// parks a lane the engine never started). This variant touches the
-    /// transcript ONLY: no status, no token/cost counters, no trace row. A
-    /// `$` command is user-initiated local shell, not agent activity.
-    ///
-    /// Folding an unknown id is a no-op — unlike `Event`, it never
-    /// auto-registers, because a lane that does not exist is precisely the
-    /// case the synthetic fallback lane already covers.
-    ShellEvent { agent: AgentId, event: AgentEvent },
-    /// One tick of a long host-side pass's live counter — `/init`'s code-graph
-    /// walk and its two embedding passes — rewritten in place by
-    /// [`crate::model::SessionModel::set_progress_line`], which carries the
-    /// reasoning for why the fold owns the rewrite.
-    ///
-    /// Deliberately neither [`Inbound::Event`] (an `AgentEvent::Text` coalesces
-    /// into one fold, which is what buried init's ✓ summaries under a wall of
-    /// repeats) nor [`Inbound::Notice`] (a notice dwells and is gone, while a
-    /// counter's final value is the part worth keeping).
-    ///
-    /// Transcript only: no status, no counters, and no trace row — the strip
-    /// keeps naming the milestone the pass is inside. Folding an unknown id is a
-    /// no-op, exactly like [`Inbound::ShellEvent`].
-    Progress { agent: AgentId, text: String },
-    /// A supervisor lifecycle transition not carried by the event stream.
-    Status { agent: AgentId, status: AgentStatus },
-    /// The dispatcher took the oldest queued prompt and handed it to an
-    /// agent. The deck's [`PromptQueue`](crate::deck::PromptQueue) is FIFO on
-    /// both sides of the channel, so this pops the front entry — the status
-    /// bar's "queued" count goes down the moment work actually starts, and a
-    /// trace row records which agent picked the prompt up.
-    PromptStarted { agent: AgentId, text: String },
-    /// The driver cancelled a turn on [`WorkspaceInput::StopAndHold`]
-    /// (double-Esc) and returned that turn's prompt to the FRONT of its
-    /// dispatch backlog. Folded as a front-insert into the deck's
-    /// [`PromptQueue`](crate::deck::PromptQueue) — the exact inverse of
-    /// [`Inbound::PromptStarted`]'s front-pop — so the queue view keeps
-    /// matching what will actually run.
-    PromptRequeued { agent: AgentId, text: String },
-    /// Reset one agent's session to its seq-0 state — a `/clear`. Folded like a
-    /// core event (it mutates the model, not just view state): the agent's
-    /// transcript is blanked, its cost/token counters and the header clock zero
-    /// out, and the progress-bar HUD returns to idle. The driver sends this on
-    /// `/clear` (alongside clearing its own LLM message history).
-    SessionReset { agent: AgentId },
-    /// A refreshed code-graph snapshot for the Graph tab. Unlike the other
-    /// variants this is **not** a folded event — the graph is an out-of-band
-    /// read-model, since a graph's structure is not in the per-session event
-    /// stream. It rides the inbound channel only because that is the
-    /// driver→deck path;
-    /// [`crate::deck_ui::ingest_inbound`] applies it straight to the view
-    /// state (`DeckUi::graph`) and the model fold ignores it. The driver
-    /// sends one after `/init` rebuilds the index so the tab reflects it
-    /// without a restart.
-    GraphSnapshot(GraphSnapshot),
-    /// How far behind the workspace's semantic index is, sent by the driver
-    /// as its background embedding pass fills it and once more when that pass
-    /// stops (#4043). Out-of-band view state like
-    /// [`Inbound::GraphSnapshot`]: applied straight to
-    /// `DeckUi::index_readiness`, ignored by the model fold. It gates one
-    /// thing — a first prompt submitted while a cold workspace is still
-    /// indexing (`deck_ui::gates::index_hold`).
-    IndexReadiness(IndexReadiness),
-    /// A refreshed slash-command vocabulary for the `/` popup. Out-of-band
-    /// view state exactly like [`Inbound::GraphSnapshot`]: applied straight
-    /// to `DeckUi::slash_commands` by [`crate::deck_ui::ingest_inbound`],
-    /// ignored by the model fold. The driver sends one after `/init` adopts
-    /// custom commands/skills so the menu reflects them without a restart.
-    SlashCommands(Vec<crate::composer::SlashCommand>),
-    /// The session's resolved triage / worker / verifier pins, sent once at
-    /// startup by the driver — which is the only side that can call
-    /// `resolve_provider` — so the statline's MODEL cell and the `/models`
-    /// dialog's standing column can name all three before any turn has run.
-    ///
-    /// Without this they stay empty until each role's first
-    /// `AgentEvent::StepUsage`, because that is the only event carrying a
-    /// role/provider/model triple. A verifier that has not been reached yet
-    /// would then read as unconfigured rather than unused.
-    ///
-    /// Folded as *configured* (`RolePin::served == false`) and drawn dim; a
-    /// later `StepUsage` replaces it with what actually served. Sending it is
-    /// optional — a driver that never does behaves exactly as before.
-    ConfiguredRoles(Vec<(crate::deck::PipelineRole, crate::deck::RolePin)>),
-    /// A deliberate mid-session re-pin: replace the named roles' pins,
-    /// served evidence included. The driver sends this after a session model
-    /// switch (`/model`, an assumed agent's `model:`), where
-    /// [`Inbound::ConfiguredRoles`]'s never-overwrite fold is the wrong
-    /// contract — the old served pin describes calls of a model that no
-    /// longer serves, and keeping it would have the statline name the wrong
-    /// model until the next `StepUsage`. Two verbs rather than a flag:
-    /// startup intent must never clobber evidence, a switch must.
-    RolePinsReset(Vec<(crate::deck::PipelineRole, crate::deck::RolePin)>),
-    /// Derived prompt-cache economics for one agent's latest model call —
-    /// dollars saved and the provider's cache TTL — computed by the
-    /// pricing-aware producer (the CLI has the model catalog; the deck does
-    /// not) and folded into the agent's [`crate::deck::AgentEntry`]. Paired
-    /// with the raw `StepUsage` the same call emits (which carries the token
-    /// counts the deck already folds): this adds only the two figures that
-    /// need list pricing / the TTL table, keeping the single savings formula
-    /// in `stella-model` and the deck free of a model-tier dependency.
-    ///
-    /// `savings_usd_delta` is this call's signed savings (negative when the
-    /// write premium outran the reads it bought — the low-hit incident), added
-    /// to the agent's running total. `ttl_secs` is the provider's prompt-cache
-    /// TTL in seconds (`0` = no prompt cache / no TTL to preserve); the deck
-    /// pairs it with the last provider-call time to render a live warmth
-    /// countdown. `is_opt_in_provider` is whether this provider only caches
-    /// behind an explicit marker (Anthropic/Bedrock/OpenRouter-Claude) —
-    /// resolved once here from `stella-model`'s cache-posture table so
-    /// [`crate::deck::AgentEntry::cache_diagnosis`] can name
-    /// `CacheCause::OptInNeverEngaged` without the deck itself needing to
-    /// know which providers require the marker.
-    CacheInsight {
-        agent: AgentId,
-        savings_usd_delta: f64,
-        ttl_secs: u64,
-        is_opt_in_provider: bool,
-    },
-    /// The installed-agents list for the Agents tab's INSTALLED AGENTS pane.
-    /// Out-of-band view state (applied straight to `DeckUi::installed` by
-    /// [`crate::deck_ui::ingest_inbound`], ignored by the model fold). The
-    /// driver — which owns the definitions on disk — sends one when the pane
-    /// asks ([`WorkspaceInput::AgentsRefresh`]) and after every save / pin /
-    /// create so the list stays live. `status`, when set, replaces the
-    /// pane's hint line (op outcomes, errors).
-    /// Which installed agent the lead is running as, after a
-    /// [`WorkspaceInput::AgentAssume`] — `None` drops back to the plain
-    /// persona. The list marks the row.
-    AgentAssumed { name: Option<String> },
-    AgentsList {
-        entries: Vec<InstalledAgentEntry>,
-        status: Option<String>,
-        /// True while an LLM-assisted agent creation is still in flight
-        /// driver-side (e.g. parked behind a running turn). The create
-        /// dialog keeps its spinner up until a list arrives with this
-        /// `false` — a queued interim snapshot must not read as done.
-        creating: bool,
-        /// The name of the agent a just-completed
-        /// [`WorkspaceInput::AgentCreate`] installed, when this snapshot is
-        /// that op's completion. The create dialog transitions into the
-        /// detail preview of exactly this entry; `None` on a failed create
-        /// (the dialog shows `status` as the error) and on every other
-        /// snapshot.
-        created: Option<String>,
-    },
-    /// A refreshed snapshot of the installed skills for the SKILLS tab. The
-    /// driver owns the skills on disk (both scopes), their enabled/version/pin
-    /// state, and the npx registry; the deck renders this read-model. Applied
-    /// straight to `DeckUi::skills` by [`crate::deck_ui::ingest_inbound`],
-    /// ignored by the model fold — same out-of-band contract as
-    /// [`Inbound::GraphSnapshot`].
-    Skills(SkillsView),
-    /// The result of a registry search (`npx skills find <query>`). Folded
-    /// into the SKILLS tab's search pane; out-of-band like [`Inbound::Skills`].
-    SkillSearch {
-        query: String,
-        hits: Vec<SkillSearchHit>,
-        status: Option<String>,
-    },
-    /// The rendered `SKILL.md` body for the ctrl+o preview overlay, fetched by
-    /// the driver (`npx skills use <id>`) for a not-yet-installed search hit.
-    /// Out-of-band like [`Inbound::SkillSearch`]; `id` lets the tab drop a
-    /// stale reply if the user closed or re-targeted the preview meanwhile.
-    SkillPreview {
-        id: String,
-        body: String,
-        status: Option<String>,
-    },
-    /// A refreshed snapshot of the configured MCP servers for the MCP tab.
-    /// Out-of-band view state exactly like [`Inbound::GraphSnapshot`]: applied
-    /// straight to `DeckUi::mcp` by [`crate::deck_ui::ingest_inbound`], ignored
-    /// by the model fold. The driver sends one at startup and after every MCP
-    /// action (install, toggle, auth, remove) so the tab reflects live state.
-    McpServers(Vec<McpServerInfo>),
-    /// The result of an MCP registry search the tab requested
-    /// ([`WorkspaceInput::McpSearch`]) — also out-of-band, applied to
-    /// `DeckUi::mcp` search results.
-    McpSearchResults(McpSearchOutcome),
-    /// One server's assembled inspector detail ([`WorkspaceInput::McpInspect`]).
-    /// Out-of-band view state like the two above. The driver may send this
-    /// **twice** for one request: once immediately from what it already knows,
-    /// and again after an optional registry lookup fills in a description — so
-    /// the inspector opens instantly and is never blocked on the network.
-    ///
-    /// Boxed: the detail carries a dozen optional strings plus the whole tool
-    /// table, and inlining it would make every `Inbound` — including the
-    /// per-token `Event` that flows thousands of times a turn — as large as
-    /// the rarest one.
-    McpDetail(Box<McpServerDetail>),
-    /// Open the help overlay. Sent by the driver when the user types `/help`
-    /// (so the slash command reaches the same rich, scrollable panel the `?`
-    /// key opens) and applied straight to `DeckUi::help_open` by
-    /// [`crate::deck_ui::ingest_inbound`]. Out-of-band view state, ignored by
-    /// the model fold — like [`Inbound::GraphSnapshot`].
-    ShowHelp,
-    /// An agent asked the driver a question and its turn is **parked** on the
-    /// answer (#4220). Raises the question overlay
-    /// ([`crate::views::question`]); the driver is holding a `QuestionResponder`
-    /// open on the other side and unblocks only on
-    /// [`WorkspaceInput::QuestionAnswered`] or its own TTL.
-    ///
-    /// Out-of-band view state like [`Inbound::ShowHelp`] — the model fold
-    /// ignores it. A parked tool call is not conversation: nothing is said
-    /// until the answer comes back, and folding a pending question into the
-    /// transcript would put a question in the history that may never have
-    /// been answered.
-    ///
-    /// Boxed for the same reason as [`Inbound::McpDetail`]: it carries a
-    /// whole batch of questions with their option lists, and the per-token
-    /// [`Inbound::Event`] must not grow to the size of the rarest variant.
-    QuestionAsked(Box<stella_protocol::QuestionRequest>),
-    /// The parked question is no longer answerable — its TTL expired, or the
-    /// turn holding it was cancelled — so take the card down.
-    ///
-    /// The half that makes the overlay safe to leave up for the full
-    /// thirty-minute TTL: without it a card would outlive its broker and
-    /// still offer to resolve, and the driver would type a considered answer
-    /// into a oneshot nobody is holding. Sent by the responder's own drop
-    /// path, so it fires on every way the wait can end rather than only the
-    /// ones somebody remembered to handle.
-    QuestionWithdrawn,
-    /// A gate demands a human yes/no before a tool call may run, and that
-    /// dispatch is **parked** on the answer (#4240). Raises the approval card
-    /// ([`crate::views::approval`]).
-    ///
-    /// Separate from [`Inbound::QuestionAsked`] because the two asks are
-    /// different decisions: this one is a yes/no over a call already chosen,
-    /// where the whole job is showing enough of the request — the tool,
-    /// whether it mutates, the gate that stopped it — to make a refusal
-    /// defensible. It rode the generic `AskUser` card until #4240, which
-    /// flattened all five fields into one line of prose.
-    ///
-    /// Out-of-band view state; the model fold ignores it. Boxed like its
-    /// siblings so the per-token [`Inbound::Event`] does not grow to the size
-    /// of the rarest variant.
-    ApprovalAsked(Box<stella_protocol::approval::ApprovalRequest>),
-    /// The parked approval is no longer answerable — its TTL expired (the
-    /// shorter of the two deadlines; `stella-cli` sets the deck's) or the turn
-    /// holding it was cancelled — so take the card down.
-    ///
-    /// The dispatch is denied on that path, never approved, so this is
-    /// strictly about not leaving a card up that offers to decide something
-    /// already decided.
-    ApprovalWithdrawn,
-    /// A launch-mark cue (see [`SplashCue`]): the driver holds the mark
-    /// open over a running init (session startup, `/init`) and
-    /// releases it when init finishes. Out-of-band view state, applied
-    /// straight to `DeckUi::splash` by [`crate::deck_ui::ingest_inbound`],
-    /// ignored by the model fold — like [`Inbound::ShowHelp`].
-    Splash(SplashCue),
-    /// A **system notification** — the deck telling the user something about
-    /// the session itself: that a previous session is resumable, what the
-    /// code-graph index pass found, that an `mcp.toml` went untrusted.
-    ///
-    /// Deliberately not [`Inbound::Event`] with an [`AgentEvent::Text`]: the
-    /// transcript is the home for agent and user messages **only**, and
-    /// routing chrome through it made the deck render machine chatter as
-    /// though the agent had said it — then kept it in the scrollback for the
-    /// rest of the session. This is out-of-band view state, applied straight
-    /// to `DeckUi::notice` by [`crate::deck_ui::ingest_inbound`] and ignored
-    /// by the model fold, like [`Inbound::ShowHelp`] and [`Inbound::Splash`].
-    ///
-    /// The deck shows these as a transient dialog ([`crate::notice`]) that any
-    /// key or mouse event dismisses and that expires on its own; it never
-    /// touches the transcript, so nothing here can be mistaken for the model
-    /// speaking.
-    Notice(String),
-    /// The plugin panels this session may draw, as of one composition of the
-    /// roster.
-    ///
-    /// **A seat is the install grant crossing into the deck.** The driver
-    /// composes the roster, drops every package whose grant is not in force,
-    /// refuses every slash name that collides with a built-in, and sends what
-    /// survives; the deck seats exactly these and can add nothing of its own.
-    /// So an ungranted plugin is never leased a rectangle — not because the
-    /// renderer declines to draw one, but because no slot for it exists.
-    ///
-    /// Sent at session open **and again whenever the roster could have
-    /// changed** (`/reload`), replacing every seat wholesale — a seat is the
-    /// grant's shadow, so a plugin retracted between two of these has to lose
-    /// its rectangle, and merging would leave it drawing from a slot nobody
-    /// renewed (#5253).
-    ///
-    /// Out-of-band view state, applied to `DeckUi::panels` and ignored by the
-    /// model fold, like [`Inbound::Notice`] above.
-    PanelsSeated {
-        /// Which composition these seats came from, counted up by the driver.
-        ///
-        /// Stamped onto every [`WorkspaceInput::PanelFrameWanted`] the deck
-        /// raises against them, so a request still in flight when the next
-        /// reseat lands names a seating that no longer exists and starts no
-        /// process. Without it a stale request reaches the driver as a bare
-        /// slot index, and an index means whatever the *route* list says it
-        /// means when it arrives — which is a different plugin's program.
-        generation: u64,
-        /// The panels that may draw, in seating order.
-        seats: Vec<PanelSeat>,
-    },
-    /// One frame a plugin drew inside its budget, for the seat at `slot`.
-    ///
-    /// The tick that produced it ran as a task off the draw path
-    /// (`stella_runtime::panel_host::ask`), and the lease it answered was
-    /// checked with `PanelLease::admits` before this was sent — so the deck
-    /// stores the frame rather than re-deciding whether it belongs.
-    PanelFrame {
-        /// Which seat drew it, as an index into [`Inbound::PanelsSeated`].
-        slot: usize,
-        /// The frame, boxed because it is much the largest thing an `Inbound`
-        /// carries and every other `Inbound` would pay for it in the enum.
-        frame: Box<stella_plugin::PanelFrame>,
-    },
-    /// A tick that produced no frame at all — the process did not start, wrote
-    /// nothing readable, or answered a lease the host had moved past.
-    ///
-    /// Distinct from [`Inbound::PanelThrottled`], which draws a tag: this one
-    /// draws nothing new and says nothing, because a plugin that failed is not
-    /// a plugin that was slow, and only the second is a fact about the frame on
-    /// screen. It exists so the deck can rearm the seat — a request with no
-    /// answer would leave the panel waiting on a tick that will never land.
-    PanelSilent {
-        /// Which seat produced nothing, as an index into
-        /// [`Inbound::PanelsSeated`].
-        slot: usize,
-    },
-    /// A tick that overran its budget: the seat keeps whatever frame is on
-    /// screen and the host tags it (SPEC 12.4).
-    PanelThrottled {
-        /// Which seat overran, as an index into [`Inbound::PanelsSeated`].
-        slot: usize,
-        /// What the tick actually cost.
-        elapsed_ms: u64,
-        /// What it was leased, so the tag reports rather than complains.
-        budget_ms: u32,
-    },
-    /// A refreshed snapshot of the **cross-process session registry** for the
-    /// SESSIONS overlay (empty-prompt `←`). Every running stella session on
-    /// this machine, grouped by [`SessionPhase`]. Out-of-band view state like
-    /// [`Inbound::GraphSnapshot`]; the driver answers
-    /// [`WorkspaceInput::SessionsRefresh`] and every archive/delete with one.
-    Sessions(Vec<SessionInfo>),
-    /// A refreshed snapshot of the persist-until-read notification store for
-    /// the inbox overlay and the footer's unread badge. The driver polls the
-    /// store (other sessions produce into it too) and pushes one whenever the
-    /// set changes. Out-of-band view state, ignored by the model fold.
-    Notifications(Vec<NotificationInfo>),
-    /// Progress from an in-flight MCP OAuth login the tab started
-    /// ([`WorkspaceInput::McpOauthLogin`]). `outcome` is `None` while running,
-    /// `Some(ok)` when finished — success triggers the tab to request a fresh
-    /// snapshot so the ⚿ oauth badge flips. Out-of-band view state.
-    McpOauthStatus {
-        server: String,
-        message: String,
-        outcome: Option<bool>,
-    },
-    /// A refreshed snapshot of the agent-engine configuration
-    /// (`settings.json` → `agent_engine_config`) for the ENGINE overlay
-    /// (the SETTINGS tab's config panel). Out-of-band view state exactly like
-    /// [`Inbound::GraphSnapshot`]: applied straight to `DeckUi` by
-    /// [`crate::deck_ui::ingest_inbound`], ignored by the model fold. The
-    /// driver sends one at startup, after every
-    /// [`WorkspaceInput::EngineConfigSave`], and on
-    /// [`WorkspaceInput::EngineConfigRefresh`]. `status`, when set,
-    /// replaces the overlay's hint line (save outcomes, errors).
-    EngineConfig {
-        state: EngineConfigState,
-        status: Option<String>,
-    },
-    /// A refreshed snapshot of the session's tool surface and the `"tools"`
-    /// switches in force, for the SETTINGS tab's tool editor. Out-of-band view
-    /// state exactly like [`Inbound::EngineConfig`].
-    ///
-    /// The driver is the only thing that can build this: MCP tools and a
-    /// customer's own registered tools exist only in the assembled session
-    /// stack, so a catalog-driven list would be a list of the tools Stella
-    /// ships — not of the tools this operator has. Sent at startup, after
-    /// every [`WorkspaceInput::ToolsSave`], and on
-    /// [`WorkspaceInput::ToolsRefresh`].
-    ToolPolicy {
-        state: ToolPolicyState,
-        status: Option<String>,
-    },
-    /// The answer to an ISSUES-tab [`WorkspaceInput::IssuesRefresh`] (and the
-    /// follow-up refresh a successful [`WorkspaceInput::IssueCreate`]
-    /// triggers): the tracker's issue list, or the error that stopped it —
-    /// including the "no tracker connected" hint the tab renders as its
-    /// empty state. Out-of-band view state like [`Inbound::McpSearchResults`];
-    /// `seq` echoes the request so the panel can drop stale replies.
-    IssuesList {
-        seq: u64,
-        outcome: Result<Vec<IssueRow>, String>,
-    },
-    /// The outcome of one ISSUES-tab mutation ([`WorkspaceInput::IssueCreate`]
-    /// / [`WorkspaceInput::IssueAct`]): a human status line on success (the
-    /// created key + url, "comment added", …) or the failure reason. `key` is
-    /// the issue acted on (the created key for a create; empty when a create
-    /// failed before a key existed). Out-of-band, seq-guarded like
-    /// [`Inbound::IssuesList`].
-    IssueActDone {
-        seq: u64,
-        key: String,
-        outcome: Result<String, String>,
-    },
-    /// The answer to an ISSUES-tab [`WorkspaceInput::IssueDraftPlan`]: the
-    /// drafted plan the start-work overlay renders, or what stopped it.
-    /// Out-of-band and seq-guarded like [`Inbound::IssuesList`]; boxed because
-    /// the draft carries the issue's sources and every task's contract, far
-    /// more than any sibling payload on this channel.
-    IssueDraft {
-        seq: u64,
-        outcome: Result<Box<StartWorkDraft>, String>,
-    },
-    /// The answer to a type-ahead [`WorkspaceInput::EntitySearch`]: the merged
-    /// hit list for the create form's Assignee/Labels popup. `query` echoes
-    /// the text searched (display only); `seq` echoes the request so the
-    /// per-keystroke stream can drop out-of-order replies — only the newest
-    /// emitted seq is ever applied.
-    EntityHits {
-        field: EntityField,
-        seq: u64,
-        query: String,
-        hits: Vec<EntityHit>,
-    },
-    /// INSPECT overlay: every model call this execution recorded a receipt for,
-    /// in wire order. Answers [`WorkspaceInput::InspectRefresh`]. An empty vec
-    /// is a real answer — an execution whose receipts predate the receipts
-    /// plane has none — and the overlay says so rather than hanging.
-    RecordedCalls(Vec<RecordedCallInfo>),
-    /// INSPECT overlay: the reconstructed context of one call, answering
-    /// [`WorkspaceInput::InspectCall`]. Boxed — the message bodies are whole
-    /// prompts, far larger than any other `Inbound` payload, and this variant
-    /// would otherwise set the size of every channel send.
-    InspectedCall(Box<InspectView>),
-    /// A failing gate has put a plan revision up for `agent` — SPEC 8.1 item 3.
-    ///
-    /// Out-of-band view-and-model state like [`Inbound::GraphSnapshot`], not
-    /// an [`Inbound::Event`]: `stella_store::plan_graph::RevisionGate` authors
-    /// it host-side from a `GateBoard` the host already evaluated, so the
-    /// journal keeps the board and the proposal is the host's reading of it.
-    /// Routing it as an event would put a host-side decision into the durable
-    /// stream, where a second producer could then write one.
-    ///
-    /// Folded by `SessionModel::propose_revision`, which files the scrollback
-    /// row and starts withholding the composer in one step.
-    RevisionProposed {
-        agent: AgentId,
-        proposal: Box<stella_protocol::RevisionProposal>,
-    },
-    /// A finished dictation's text ([`WorkspaceInput::VoiceStop`]'s answer).
-    /// Ingest pastes it at the composer's cursor and returns
-    /// [`crate::voice::VoiceUi`] to rest — never a model fold: a dictation
-    /// is keyboard input that arrived by another route.
-    VoiceTranscript { text: String },
-    /// A dictation that produced no text (no recorder, no credential, a
-    /// refused transcription request). Ingest returns the voice state to
-    /// rest and raises `reason` as a transient notice.
-    VoiceFailed { reason: String },
-    /// Dictation was switched on, off, or between modes by `/voice`, which
-    /// runs on the driver side and has just persisted the same values to
-    /// settings.
-    ///
-    /// Without this the deck would keep the enablement and mode it read at
-    /// startup, so `/voice tap` would report success and change nothing until
-    /// the next session — and "first use is one command" is the whole point
-    /// of the command (#5347). Any gesture in flight is returned to rest:
-    /// the keys mean something different the instant the mode changes, and a
-    /// half-finished hold folded by tap-mode rules is a capture nobody can
-    /// stop.
-    VoiceConfig {
-        enabled: bool,
-        mode: crate::voice::VoiceMode,
-    },
-}
-
-// The ISSUES tab's read models. Re-exported rather than moved behind a
-// path, so no consumer has to learn where they went.
-mod issues;
-pub use issues::{EntityField, EntityHit, IssueAction, IssueRow, LinkedWork};
-
-/// The session-registry lifecycle phase, exactly the grouping the SESSIONS
-/// overlay shows. A TUI-local mirror of `stella-store`'s `SessionStatus`
-/// (the deck never links the store crate; the driver maps one to the other).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum SessionPhase {
-    InProgress,
-    NeedsInput,
-    /// Set aside with its durable state intact — quit (or switched away
-    /// from) with work still pending; the first thing resume looks for.
-    Paused,
-    Cancelled,
-    /// Ended itself by policy (stuck-loop escalation, step cap, enforced
-    /// budget, ended scope review) — a deliberate ending beside `Cancelled`,
-    /// never an error (#1653).
-    Stopped,
-    Complete,
-    Archived,
-    Error,
-}
-
-impl SessionPhase {
-    /// Display/grouping order: attention-worthy first.
-    pub const ALL: [SessionPhase; 8] = [
-        SessionPhase::InProgress,
-        SessionPhase::NeedsInput,
-        SessionPhase::Paused,
-        SessionPhase::Cancelled,
-        SessionPhase::Stopped,
-        SessionPhase::Complete,
-        SessionPhase::Archived,
-        SessionPhase::Error,
-    ];
-
-    pub fn label(self) -> &'static str {
-        match self {
-            SessionPhase::InProgress => "In Progress",
-            SessionPhase::NeedsInput => "Needs Input",
-            SessionPhase::Paused => "Paused",
-            SessionPhase::Cancelled => "Cancelled",
-            SessionPhase::Stopped => "Stopped by policy",
-            SessionPhase::Complete => "Complete",
-            SessionPhase::Archived => "Archived",
-            SessionPhase::Error => "Error",
-        }
-    }
-}
-
-/// One row of the SESSIONS overlay — a running (or finished) stella session
-/// from the machine-wide registry, with the human title and work summary the
-/// registry holds.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct SessionInfo {
-    /// Registry id (`ses-…`), the archive/delete handle.
-    pub id: String,
-    /// Human title: `<workspace basename>: <first prompt…>`.
-    pub title: String,
-    /// What work is involved — the latest prompt/goal, truncated.
-    pub summary: String,
-    /// One sentence on what the session did, written by a model from its
-    /// prompts (`stella-cli`'s `sessions_view`); `None` until it has been.
-    pub description: Option<String>,
-    /// Workspace path (dimmed detail line).
-    pub workspace: String,
-    pub phase: SessionPhase,
-    pub started_ms: u64,
-    pub updated_ms: u64,
-    /// True for the record of THIS deck process (rendered with a marker and
-    /// protected from delete).
-    pub mine: bool,
-    /// True when the session can be reopened here: no live process owns it,
-    /// it belongs to this deck's workspace, and its durable state (journal /
-    /// history) is on disk. `⏎` on such a row sends
-    /// [`WorkspaceInput::SessionResume`].
-    pub resumable: bool,
-    /// Turns recorded in the store for this session.
-    pub turns: u32,
-    /// Spend across those turns, in micro-dollars — integral so the row stays
-    /// `Eq`, and six decimals is what the store keeps.
-    pub spend_micros: u64,
-    /// The model the latest turn ran on.
-    pub model: Option<String>,
-}
-
-/// One persist-until-read notification as the inbox overlay lists it. A
-/// mirror of `stella-store`'s `Notification` minus storage concerns.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct NotificationInfo {
-    pub id: String,
-    pub title: String,
-    pub body: String,
-    /// Origin hint (session id, server name); may be empty.
-    pub source: String,
-    pub created_ms: u64,
-    pub read: bool,
-    /// The session this notification is about, when it has one — what lets
-    /// the inbox's `Enter` open the session (replaying it if it is no longer
-    /// live) via [`WorkspaceInput::SessionOpen`].
-    pub session_id: Option<String>,
-}
-
-/// Driver → deck cues for the launch mark ([`crate::splash`]).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum SplashCue {
-    /// Show the mark **held** open over a running init: it stands until
-    /// `Release`. Ignored on `--no-anim` sessions (a static frame is their
-    /// contract).
-    Replay,
-    /// Init finished — hand the frame straight back to the deck. A no-op if
-    /// no held mark is showing.
-    Release,
-}
-
-/// Which config level an installed agent definition lives at.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum AgentScope {
-    /// The workspace's `.stella/agents/` directory.
-    Project,
-    /// The user's `~/.stella/agents/` directory.
-    User,
-}
-
-impl AgentScope {
-    pub fn label(self) -> &'static str {
-        match self {
-            AgentScope::Project => "project",
-            AgentScope::User => "user",
-        }
-    }
-}
-
-/// One selectable version of an installed agent (the version picker's row).
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct AgentVersionInfo {
-    /// 1-based version number.
-    pub version: u32,
-    /// Short display label (e.g. the version file's modification time), or
-    /// empty when none was available.
-    pub label: String,
-}
-
-/// One installed agent as the Agents tab's INSTALLED AGENTS pane lists it
-/// (see [`Inbound::AgentsList`]). Decoupled from `stella-core`'s `AgentDef`
-/// so the TUI crate stays independent of the extensions engine — the driver
-/// maps one to the other and adds the version/scope bookkeeping.
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct InstalledAgentEntry {
-    /// The agent's loaded definition name.
-    pub name: String,
-    /// One-line description shown beside the name.
-    pub description: String,
-    /// The toolbelt grant from the definition's `tools:` frontmatter.
-    /// `None` = the definition does not restrict tools — the agent gets
-    /// **all** tools, and the pane says so honestly.
-    pub tools: Option<Vec<String>>,
-    /// Which config level the definition lives at.
-    pub scope: AgentScope,
-    /// The definition file the loader reads (display/provenance).
-    pub source_path: String,
-    /// The pinned (active) version — 1 for a never-versioned agent.
-    pub version: u32,
-    /// Every version on disk, oldest first. Always contains `version`.
-    pub versions: Vec<AgentVersionInfo>,
-    /// The pinned version's full file content (frontmatter + body) — what
-    /// the editor loads.
-    pub content: String,
-}
-
-/// How hard a bang-persistence sigil asks the record it saves to steer.
-///
-/// The two sigils differ by this field alone, so one write path serves both.
-/// `stella_tui::deck_ui::sigil` parses them.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum KeepStrength {
-    /// `!!` — a `should` record. It steers, and a later explicit instruction
-    /// can still override it.
-    Guidance,
-    /// `!!!` — a `must` record at the precedence ceiling. It holds until
-    /// somebody retires it.
-    Rule,
-}
-
-/// What the deck sends back to the caller / engine. The single-session
-/// [`UserInput`] is wrapped with the target agent; new verbs cover the deck's
-/// workspace-level affordances (queueing, agent control, quit).
-#[derive(Clone, Debug, PartialEq)]
-pub enum WorkspaceInput {
-    /// Route a `UserInput` (prompt, scope decision, ask-user answer) to one agent.
-    ToAgent { agent: AgentId, input: UserInput },
-    /// Queue a brand-new prompt without blocking on any busy agent — the
-    /// router picks the model/agent. The deck never gates input on agent state.
-    Enqueue { text: String },
-    /// Queue a prompt at the FRONT: the first submission after a
-    /// [`WorkspaceInput::StopAndHold`]. The deck sends this instead of
-    /// [`WorkspaceInput::Enqueue`] while dispatch is held, so the user's new
-    /// prompt runs before the prompt the hold returned to the queue — and
-    /// receiving it is what releases the hold.
-    EnqueueFront { text: String },
-    /// Queue a prompt as the lead's NEXT turn, behind whatever is already
-    /// waiting, and never fork it to a sidecar lane. The deck sends this for
-    /// a plain prompt typed while the lead is running under the default
-    /// [`MidTurnPrompt::Queue`](crate::deck_ui::MidTurnPrompt) policy, so the
-    /// backlog is exactly what an Esc then steers into the running turn
-    /// ([`WorkspaceInput::Steer`]). At rest it dispatches like
-    /// [`WorkspaceInput::Enqueue`].
-    ///
-    /// A third verb rather than a flag on `Enqueue` because the driver's
-    /// mid-turn arm routes a plain `Enqueue` to a sidecar (`route_mid_turn`)
-    /// and `EnqueueFront` jumps the backlog; neither is "wait your turn".
-    EnqueueNext { text: String },
-    /// Remove one not-yet-dispatched prompt from the queue (0 = oldest). The
-    /// deck's queue editor sends this for `ctrl+x` delete and for pulling a
-    /// prompt back into the composer to edit it.
-    QueueRemove { index: usize },
-    /// Drop every not-yet-dispatched prompt (the deck confirms with a second
-    /// `ctrl+d` before sending this).
-    QueueClear,
-    /// `/clear`, as an event rather than a queued prompt: reset the session to
-    /// its seq-0 state NOW — even mid-turn, where the driver cancels the
-    /// in-flight turn first. The deck sends this instead of enqueueing the
-    /// text, because a reset that waits behind the backlog is not a reset (the
-    /// user watched `/clear` sit in the queue popup as "pending"). The driver
-    /// answers with [`Inbound::SessionReset`] once its own history is blanked;
-    /// the deck's only optimistic mirror is dropping its queue view, since a
-    /// session reset to seq-0 has no backlog by definition.
-    SessionClear,
-    /// A slash command whose declared class is **queue-free**
-    /// ([`crate::composer::SlashCommand::sideband`]): the deck sends this
-    /// instead of [`WorkspaceInput::Enqueue`] so the command executes at
-    /// once — mid-turn included — beside the prompt queue rather than in it.
-    /// The driver runs it without touching the in-flight turn, the backlog,
-    /// or the conversation, and answers into the transcript over
-    /// [`Inbound::ShellEvent`] (no status flip — the turn's own events keep
-    /// telling the truth about the turn). `/export` is the canonical case:
-    /// it used to sit in the queue popup as a "pending prompt" until the
-    /// lead finished. The turn-coupled exceptions (`/clear`, `/init`,
-    /// `/reload`, custom expansions) never ride this route.
-    Command { text: String },
-    /// The agents page's "describe a task for a new session": start a NEW
-    /// sub-agent lane running `text`, whatever the lead is doing. The driver
-    /// answers with the ordinary lane birth — [`Inbound::Register`], a
-    /// [`Inbound::PromptStarted`] on that lane, a lead-transcript notice —
-    /// or, when every worker slot is taken, a [`Inbound::ShellEvent`] notice
-    /// saying so. Distinct from [`WorkspaceInput::Enqueue`], whose idle-arm
-    /// meaning is "the lead's next turn": this one is a lane in both arms.
-    SpawnLane { text: String },
-    /// How the driver settled the parked question the overlay was showing
-    /// (#4220) — the return leg of [`Inbound::QuestionAsked`].
-    ///
-    /// The whole outcome travels, not just the answers: `Deferred` ("the
-    /// options are the wrong shape, let's talk") and `Declined` (cancelled)
-    /// are real answers the asking model acts on differently, and collapsing
-    /// them into "no answers" would tell it the same thing three ways.
-    ///
-    /// Boxed to match [`Inbound::QuestionAsked`]: the answer set carries a
-    /// note and free text per question.
-    QuestionAnswered(Box<stella_protocol::QuestionOutcome>),
-    /// How the driver decided the parked approval the card was showing
-    /// (#4240) — the return leg of [`Inbound::ApprovalAsked`].
-    ///
-    /// A `Deny` carries the driver's own words when they gave any, and the
-    /// broker forwards them to the model verbatim: "no, use the staging
-    /// bucket" is a redirection the turn can act on, where a bare refusal is
-    /// a wall it has to guess its way around.
-    ApprovalAnswered(Box<stella_protocol::approval::ApprovalResponse>),
-    /// Pause / resume / stop / restart a specific agent.
-    Control {
-        agent: AgentId,
-        control: AgentControl,
-    },
-    /// Double-Esc: cancel `agent`'s in-flight turn, return that turn's
-    /// prompt to the FRONT of the queue, and HOLD dispatch until the user's
-    /// next submission — "full stop; what I type next runs first". A single
-    /// Esc is the plain [`AgentControl::Stop`]: the lead SOFT-stops at the
-    /// next step boundary (completed steps kept); worker lanes cancel
-    /// immediately, and the next queued prompt dispatches automatically.
-    StopAndHold { agent: AgentId },
-    /// Esc with something queued: inject **every** waiting prompt into
-    /// `agent`'s running turn at its next step boundary, in order, and keep
-    /// the turn going. See [`steering`] for why this is not a cancel, and why
-    /// the whole queue travels in one message.
-    Steer { agent: AgentId, texts: Vec<String> },
-    /// The composer's red mode: soft-stop `agent`'s running turn (completed
-    /// steps kept, exactly the first-Esc stop) and run `texts` next, ahead of
-    /// the parked backlog. With no running turn it degrades to "run this
-    /// now", and at a worker lane it lands as a steer — a lane's next prompt
-    /// is the lead's, so there is nothing for the words to outrank there.
-    ///
-    /// `keep` carries the bang-persistence sigils (`!!`, `!!!`): the same
-    /// stop, plus the words saved as a context record. It rides this message
-    /// rather than a second one so the stop and the save cannot separate — a
-    /// save whose stop was dropped, or a stop whose save was, is a failure the
-    /// person who typed it never sees. `None` at every other door.
-    Interrupt {
-        agent: AgentId,
-        texts: Vec<String>,
-        keep: Option<KeepStrength>,
-    },
-    /// The composer's broadcast address (`>@all …`, `>>> @agents !!! …`):
-    /// send the words to the live sessions on this machine over their
-    /// whistle sockets. [`Broadcast`] holds who hears it, how deep it
-    /// reaches, whether their turns stop, and what is kept. Queue-free like a
-    /// sideband command: it is about other sessions, so it never waits on
-    /// this one's turn. The driver answers with a chrome note of per-session
-    /// outcomes. A plain `>@all` leaves this session alone; a `>>> @agents`
-    /// interrupt reaches it too, in process, because the room includes the
-    /// person who spoke.
-    Whistle(Broadcast),
-    /// Re-root the Graph tab on `file`: the deck's file picker sends this when
-    /// the user selects a file, and the driver answers with a fresh
-    /// [`Inbound::GraphSnapshot`] centered on it (the same out-of-band refresh
-    /// path `/init` uses). `stella-tui` cannot query the graph store itself, so
-    /// re-rooting is a round-trip rather than a local recompute — the picker
-    /// only knows the file *names* (shipped in [`GraphSnapshot::files`]), never
-    /// their neighborhoods. `file` is a root-relative path from that list.
-    FocusGraphFile { file: String },
-    /// Re-root the Graph tab on a free-form query — a symbol name, or any
-    /// text the driver's index can resolve to definitions. Sent by the tab's
-    /// `q` box; answered, like [`Self::FocusGraphFile`], with a fresh
-    /// [`Inbound::GraphSnapshot`], whose
-    /// [`query`](GraphSnapshot::query) echoes `text` back so the bar shows
-    /// what the neighborhood on screen actually answers (#4335).
-    ///
-    /// A separate verb rather than a mode flag on `FocusGraphFile`: the two
-    /// resolve differently on the driver side (a path lookup versus a symbol
-    /// lookup), and a query that matches nothing has to be able to say so
-    /// without being mistaken for a missing file.
-    GraphQuery { text: String },
-    /// The INSTALLED AGENTS pane opened (or wants a reload): enumerate the
-    /// agent definitions installed at both scopes and answer with
-    /// [`Inbound::AgentsList`].
-    AgentsRefresh,
-    /// Save an edited agent definition as a NEW version and pin it — the
-    /// prior version is preserved on disk (see `stella-cli`'s
-    /// `agents_installed` module for the on-disk scheme). The driver
-    /// answers with a fresh [`Inbound::AgentsList`].
-    AgentSave {
-        name: String,
-        scope: AgentScope,
-        content: String,
-    },
-    /// Re-pin an existing version WITHOUT editing — the version count never
-    /// changes on a pin (increments happen only on [`WorkspaceInput::AgentSave`]).
-    AgentPin {
-        name: String,
-        scope: AgentScope,
-        version: u32,
-    },
-    /// INSTALLED AGENTS `x x`: delete the definition — its canonical file
-    /// and its archived versions. Answered with a fresh
-    /// [`Inbound::AgentsList`].
-    AgentDelete { name: String, scope: AgentScope },
-    /// INSTALLED AGENTS `a`, or the `/agent` picker: the lead assumes this
-    /// agent's identity for the rest of the session — its definition becomes
-    /// the persona the system prompt carries, its `tools:` grant narrows the
-    /// session tool policy, and its declared `model:` (if any) switches the
-    /// session model, all from the next turn on. Between turns only;
-    /// mid-turn the driver answers with a transcript notice. Answered with
-    /// [`Inbound::AgentAssumed`].
-    AgentAssume { name: String, scope: AgentScope },
-    /// The `/model` picker's selection (or `/model <spec>` typed): switch
-    /// THIS session's model to `spec` (`provider/slug`). Session-only — the
-    /// driver never writes settings for it, so future sessions keep the
-    /// configured default. Between turns only, like
-    /// [`WorkspaceInput::AgentAssume`]; the driver answers with a fresh
-    /// [`Inbound::Register`] + [`Inbound::ConfiguredRoles`] and a chrome
-    /// note (on refusal, the note alone).
-    ModelOverride { spec: String },
-    /// Create a new agent from a short description with LLM assistance: the
-    /// driver drafts the definition through the session's provider, installs
-    /// it at `scope`, and answers with a fresh [`Inbound::AgentsList`].
-    AgentCreate {
-        description: String,
-        scope: AgentScope,
-    },
-    /// A SKILLS-tab operation (list / enable / uninstall / search / install /
-    /// create / edit / pin). The driver owns the skills on disk + npx + model
-    /// and answers with a refreshed [`Inbound::Skills`] / [`Inbound::SkillSearch`].
-    Skill(SkillOp),
-    /// MCP tab: flip a configured server's session enable/disable state. The
-    /// driver toggles the shared disabled-servers set (hiding/showing the
-    /// server's tools on the next model call) and pushes a fresh
-    /// [`Inbound::McpServers`] snapshot.
-    McpToggle { name: String },
-    /// MCP tab: search the configured registry for `query`. The driver runs
-    /// the async search and replies with [`Inbound::McpSearchResults`].
-    McpSearch { query: String },
-    /// MCP tab: install the registry server named `name` into `.stella/mcp.toml`
-    /// (then refresh the snapshot).
-    McpInstall { name: String },
-    /// MCP tab: remove a configured server from `.stella/mcp.toml`.
-    McpRemove { name: String },
-    /// MCP tab: set an auth credential (env var for stdio, header for http) on a
-    /// configured server. The value is a [`Secret`] — its `Debug` is redacted,
-    /// so it never reaches the deck's debug log.
-    McpAuth {
-        server: String,
-        field: String,
-        value: Secret,
-    },
-    /// MCP tab: record the operator's capability grant for a configured
-    /// server, after the first-enable handshake showed what it declares
-    /// (SPEC §9.3).
-    ///
-    /// The driver writes `granted = true` into `.stella/mcp.toml`, adds the
-    /// name to the session's shared grant set — so the server becomes usable
-    /// in the session it was reviewed in, without a reconnect — enables it,
-    /// and pushes a fresh [`Inbound::McpServers`] snapshot.
-    ///
-    /// Only ever sent for a grant. There is no `McpGrant { granted: false }`:
-    /// declining is not answering, the gate's default is withheld, and a
-    /// parameter that selects between granting and revoking would be two verbs
-    /// wearing one name — the single-purpose rule AGENTS.md #9 states.
-    McpGrant { name: String },
-    /// MCP tab: rebuild and re-push the [`Inbound::McpServers`] snapshot.
-    McpRefresh,
-    /// MCP tab: assemble the ctrl+o inspector detail for one configured server
-    /// and answer with [`Inbound::McpDetail`].
-    ///
-    /// `lookup` asks the driver to consult the registry when the server has no
-    /// recorded description, backfilling `mcp.toml` with what it finds. Opt-in
-    /// per request rather than automatic: the registry is a third-party
-    /// service, and merely looking at a tab must not talk to one.
-    McpInspect { name: String, lookup: bool },
-    /// MCP tab: start the browser OAuth login for a configured **http**
-    /// server. The driver runs the flow in the background and streams
-    /// [`Inbound::McpOauthStatus`] updates (including the authorize URL).
-    McpOauthLogin { server: String },
-    /// SESSIONS overlay opened (or `r`): read the machine-wide session
-    /// registry and answer with [`Inbound::Sessions`].
-    SessionsRefresh,
-    /// SESSIONS overlay / inbox: open a session in a replay lane. The driver
-    /// loads the session's persisted event journal from the store (linked by
-    /// `session_id` since store schema v8), then answers with a normal
-    /// [`Inbound::Register`] (a `replay:<id>` lane) followed by every
-    /// persisted event as ordinary [`Inbound::Event`]s and a terminal
-    /// [`Inbound::Status`] — replay IS the fold, so a session dead for 12
-    /// hours reconstructs to exactly the state it died in, with no second
-    /// rendering path.
-    SessionOpen { id: String },
-    /// SESSIONS overlay: tuck a session record away (status → Archived).
-    /// Answered with a fresh [`Inbound::Sessions`].
-    SessionArchive { id: String },
-    /// SESSIONS overlay: delete a session record from the registry.
-    /// Answered with a fresh [`Inbound::Sessions`].
-    SessionDelete { id: String },
-    /// SESSIONS overlay: reopen a resumable session (⏎ on a
-    /// [`SessionInfo::resumable`] row) — the deck-native "navigate back
-    /// into a session". The driver parks the current session (its durable
-    /// state is already on disk; its record flips to Paused), replays the
-    /// chosen session's journal through the fold, restores its conversation
-    /// and prompt backlog, and re-owns its registry record. Only serviced
-    /// between turns — mid-turn the driver answers with a transcript notice
-    /// instead of tearing down live work.
-    SessionResume { id: String },
-    /// SESSIONS overlay `n`: park this session and open a fresh, empty one
-    /// — the same hand-over [`WorkspaceInput::SessionResume`] performs, with
-    /// a new record in place of a stored one. Between turns only; mid-turn
-    /// the driver answers with a transcript notice.
-    SessionNew,
-    /// Inbox overlay: mark one notification read (it may then be pruned —
-    /// "persists until read" is the store's contract). Answered with a fresh
-    /// [`Inbound::Notifications`].
-    NotificationRead { id: String },
-    /// Inbox overlay: mark everything read.
-    NotificationsReadAll,
-    /// ENGINE overlay: persist the edited agent-engine configuration into
-    /// `settings.json` at `scope` (project `.stella/settings.json` or the
-    /// user's `~/.stella/settings.json`). The driver writes the
-    /// `agent_engine_config` object — preserving every other key in the
-    /// file — and answers with a fresh [`Inbound::EngineConfig`] carrying
-    /// the save outcome in `status`. Saved config applies to runs started
-    /// afterwards; in-flight turns keep their resolved models.
-    EngineConfigSave {
-        state: EngineConfigState,
-        scope: AgentScope,
-    },
-    /// ENGINE overlay opened (or wants a reload): re-read the settings
-    /// scope chain and answer with a fresh [`Inbound::EngineConfig`].
-    EngineConfigRefresh,
-    /// One seated panel wants its next frame, against the rectangle the deck
-    /// has just drawn it into.
-    ///
-    /// **This is what makes the panel loop run off the draw path** (SPEC 12.4).
-    /// The deck cannot ask a plugin for a frame itself — a repaint that waits
-    /// on somebody else's process is a frozen terminal — so it asks the
-    /// driver, which spawns the exchange as a task and answers with
-    /// [`Inbound::PanelFrame`], [`Inbound::PanelThrottled`] or
-    /// [`Inbound::PanelSilent`]. Exactly one of the three comes back for every
-    /// one of these, which is how the seat knows to ask again.
-    ///
-    /// The geometry rides on the request because the lease is the deck's to
-    /// state: a plugin addresses cells inside a rectangle the host measured,
-    /// and the driver has never seen the terminal.
-    PanelFrameWanted {
-        /// The seating this request belongs to, echoed from
-        /// [`Inbound::PanelsSeated`].
-        ///
-        /// The driver refuses a request naming any other, which is what stops
-        /// a request minted before a reseat from starting the plugin that now
-        /// holds its slot. `PanelSlot::settle` is the same rule on the return
-        /// leg and does not subsume this one: that one keeps a stale *frame*
-        /// off the screen, and by the time a frame exists somebody's program
-        /// has already run.
-        generation: u64,
-        /// Which seat, as an index into [`Inbound::PanelsSeated`].
-        slot: usize,
-        /// The host's counter for this frame, echoed on the frame that
-        /// answers so a late one is discardable.
-        tick: u64,
-        /// The columns the deck leased it, inside the host's chrome.
-        cols: u16,
-        /// The rows the deck leased it, inside the host's chrome.
-        rows: u16,
-    },
-    /// TOOLS panel: persist the operator's tool switches into `settings.json`
-    /// at `scope`, answered with a fresh [`Inbound::ToolPolicy`].
-    ///
-    /// `switches` carries only the keys the panel actually changed — a tool
-    /// name, a group name, or `"*"` — and the driver merges them into that
-    /// scope's own `"tools"` object rather than replacing it. Sending the
-    /// whole merged map instead would copy the OTHER scopes' switches into
-    /// the file being written and freeze them there.
-    ToolsSave {
-        switches: BTreeMap<String, bool>,
-        scope: AgentScope,
-    },
-    /// TOOLS panel opened (or wants a reload): re-read the settings scope
-    /// chain, re-enumerate the session's tools, and answer with a fresh
-    /// [`Inbound::ToolPolicy`].
-    ToolsRefresh,
-    /// ISSUES tab: list (or tracker-search) issues. `query`/`state` are the
-    /// tracker-side filters; the driver answers with [`Inbound::IssuesList`]
-    /// echoing `seq` so stale replies can be dropped.
-    ///
-    /// `page` is 0-based and is what the tab's `]`/`[` move. It rides the
-    /// request rather than being held driver-side because the panel is the
-    /// only thing that knows which page the human is looking at — and a
-    /// paging key that could not say so would re-fetch page one under a
-    /// notice claiming otherwise.
-    IssuesRefresh {
-        query: Option<String>,
-        state: Option<String>,
-        page: usize,
-        seq: u64,
-    },
-    /// ISSUES tab: create an issue from the `n` form. The driver answers
-    /// with [`Inbound::IssueActDone`] (the created key + url on success) and
-    /// then a fresh [`Inbound::IssuesList`] under the same `seq`.
-    IssueCreate {
-        title: String,
-        body: String,
-        labels: Vec<String>,
-        assignee: Option<String>,
-        seq: u64,
-    },
-    /// ISSUES tab: act on one existing issue (comment / set-status / start
-    /// work). Answered with [`Inbound::IssueActDone`].
-    IssueAct {
-        key: String,
-        action: IssueAction,
-        seq: u64,
-    },
-    /// ISSUES tab: draft a plan for one issue (`w`), from the issue's text,
-    /// the files the code graph couples to it, and the memory RULEs that
-    /// apply. Answered with [`Inbound::IssueDraft`].
-    ///
-    /// Reads only. The draft is what the human approves, so nothing on this
-    /// path may create a branch, take a claim, or spend a model call — see
-    /// [`WorkspaceInput::IssueStartWork`], which is the one request that does.
-    IssueDraftPlan { key: String, seq: u64 },
-    /// ISSUES tab: approve a drafted plan and start work on it (`a` in the
-    /// start-work overlay). `tasks` is the draft's task list **after** the
-    /// human's edits, in plan order.
-    ///
-    /// This is the only ISSUES-tab request that changes anything outside the
-    /// tracker: the driver takes the workspace's `issue:<n>` dispatch claim,
-    /// opens the branch under it, and authors the plan's first revision.
-    /// Answered with [`Inbound::IssueActDone`].
-    IssueStartWork {
-        key: String,
-        tasks: Vec<String>,
-        seq: u64,
-    },
-    /// ISSUES tab: one per-keystroke type-ahead query from the create form's
-    /// Assignee/Labels field. Answered with [`Inbound::EntityHits`] echoing
-    /// `seq` — the panel keeps only the newest.
-    EntitySearch {
-        field: EntityField,
-        query: String,
-        seq: u64,
-    },
-    /// INSPECT overlay: list the model calls this execution recorded, so a
-    /// human can pick one. Answered with [`Inbound::RecordedCalls`].
-    InspectRefresh,
-    /// INSPECT overlay: reconstruct one call's context. Answered with
-    /// [`Inbound::InspectedCall`]. A blocking SQLite read on the driver side —
-    /// it replays the block registry and the event journal — so it is served
-    /// off the pump, like [`WorkspaceInput::FocusGraphFile`].
-    InspectCall {
-        turn_instance: u32,
-        step: u64,
-        call_seq: u64,
-    },
-    /// Plan card (`/plan`), `x`: ask the driver to skip one still-open step on
-    /// `agent`'s plan. Send-and-forget — the row's state changes only when
-    /// the driver's next `TaskUpdate` snapshot folds back, so the card can
-    /// never show a skip the engine refused.
-    TaskSkip { agent: AgentId, id: String },
-    /// `u` on a highlighted delete event: restore what that one `delete_file`
-    /// call removed, from git — the row's `· git-backed · u undo` affordance
-    /// (SPEC 6.3, SPEC 11). `paths` carries the call's targets, several for a
-    /// batch delete, so the input scopes exactly one event. The driver answers
-    /// with [`Inbound::Notice`] naming what was restored or why it could not
-    /// be — an untracked file has no git reading to restore from.
-    UndoDelete { paths: Vec<String> },
-    /// `x` on a highlighted memory row: tombstone that memory so it stops
-    /// steering the agent and the reflection loop stops re-learning it — the
-    /// row's `· x reject` affordance (SPEC 6.3).
-    ///
-    /// `text` rides along because the tombstone is content-addressed as well
-    /// as id-addressed: the loop re-mines paraphrases, so a rejection stored
-    /// against the id alone would be undone by the next turn that re-learned
-    /// the same lesson under a new one. The driver answers with
-    /// [`Inbound::Notice`].
-    RejectMemory { memory_id: String, text: String },
-    /// `e edit` on a logged memory row (SPEC 6.3): a revision on the memory's
-    /// lineage, so the old words stop being served and the id does not change.
-    ///
-    /// `text` is the replacement only. [`Self::RejectMemory`] carries the old
-    /// words too, because a tombstone has to recognise a re-mined paraphrase;
-    /// an edit names a lineage the id already identifies.
-    EditMemory { memory_id: String, text: String },
-    /// `r` on a highlighted gate board: ask for one gate to be evaluated
-    /// again — SPEC 8.1's `r rerun gate`.
-    ///
-    /// A *request*, and the driver's answer is an [`Inbound::Notice`] either
-    /// way. Verification belongs to the plugin that reported the evidence and
-    /// stella never re-runs a gate itself (AGENTS.md's opening), so a door with
-    /// no verification plugin bound has no gate to re-request and says so
-    /// rather than pretending. That is the whole of the contract here: the key
-    /// reaches a driver, the driver answers in words, and no surface renders a
-    /// re-run that did not happen.
-    RerunGate {
-        /// The gate's name, as its row states it.
-        gate: String,
-    },
-    /// `a` on a highlighted revision proposal: make the plan revision the
-    /// failing gate put up — SPEC 8.1's `a approve r4`.
-    ///
-    /// The whole proposal travels rather than its subject alone, because the
-    /// driver re-checks it before writing: `RevisionGate::approve` refuses a
-    /// proposal whose revision number the plan has moved past, and a subject
-    /// on its own carries no number to check.
-    ///
-    /// `e edit` is absent from this enum on purpose: it is answered inside the
-    /// deck, which hands the subject to the composer and stands the proposal
-    /// down, so a round trip would be a message whose only effect is latency.
-    ApproveRevision {
-        agent: AgentId,
-        proposal: Box<stella_protocol::RevisionProposal>,
-    },
-    /// `x` on a highlighted revision proposal: drop it — SPEC 8.1's
-    /// `x dismiss`.
-    ///
-    /// It writes no revision, and it still has to travel. A running turn
-    /// withholds its tool calls while the change stands, and only the driver
-    /// can reach the gate that withholds them. A dismissal the deck answered
-    /// alone would clear the card and leave the turn refused.
-    DismissRevision {
-        agent: AgentId,
-        proposal: Box<stella_protocol::RevisionProposal>,
-    },
-    /// Plan card (`/plan`), post-approval `e`: ask the driver to open a
-    /// scope-change proposal for `agent`'s locked scope. The deck never edits
-    /// scope locally — a granted change arrives back as a fresh
-    /// `ScopeReview` fold.
-    ScopeChangeRequest { agent: AgentId },
-    /// Budget editor (`/budget`): set (or with `None` clear) the session
-    /// spend cap. The deck renders only the cap the budget stream folds back
-    /// (`AgentEvent::BudgetTick::session_limit_usd`), so an ignored or
-    /// clamped request never shows a cap that is not in force.
-    SetBudget { limit_usd: Option<f64> },
-    /// Push-to-talk crossed its warmup: start capturing microphone audio.
-    /// The driver answers the eventual [`WorkspaceInput::VoiceStop`] with
-    /// [`Inbound::VoiceTranscript`] or [`Inbound::VoiceFailed`]; a start it
-    /// cannot honour answers [`Inbound::VoiceFailed`] immediately.
-    VoiceStart,
-    /// The held key was released: stop capturing and transcribe what was
-    /// heard.
-    VoiceStop,
-    /// Abandon the capture without transcribing (Esc while listening).
-    VoiceCancel,
-    /// Tear down the deck.
-    Quit,
-}
-
-/// The `/models` role table: the open role vocabulary and its render order.
-pub mod roles;
-
-mod engine_config;
-/// The MCP tab's read models (rows, search results, inspector detail).
-mod mcp;
-mod skills;
-/// Why Esc steers rather than cancels — doc-only.
-pub mod steering;
-mod tool_policy;
-pub use engine_config::{EngineAgentState, EngineConfigState, EngineRole, RoleWiringRow, SeatRow};
-pub use mcp::{
-    GRAPH_SERVER, McpLiveIdentity, McpLookupState, McpSearchItem, McpSearchOutcome,
-    McpServerDetail, McpServerInfo, McpSignature, McpSourceTier, McpToolRow,
-};
-pub use roles::{RoleTableEntry, role_table};
-pub use skills::{
-    LearnedProvenance, LearnedSource, RejectedSkillRow, SkillOp, SkillRow, SkillScope,
-    SkillSearchHit, SkillsView,
-};
-pub use tool_policy::{ToolDenial, ToolPolicyState, ToolRow, ToolScope};
-
-/// A secret string whose `Debug` is redacted, so it can ride the deck's input
-/// channel (and any debug log of it) without leaking. The value is readable
-/// only via [`Secret::reveal`], used solely to write the credential to config.
-///
-/// The plaintext is wiped on drop, so a credential typed into the MCP auth
-/// prompt does not sit in freed heap for the rest of the deck session. The
-/// wipe covers the buffer this value owns at drop time and nothing else —
-/// notably not an earlier allocation abandoned when the underlying `String`
-/// grew, which no `Drop` impl can reach.
-#[derive(Clone, PartialEq, Eq)]
-pub struct Secret(String);
-
-impl Secret {
-    pub fn new(value: impl Into<String>) -> Self {
-        Secret(value.into())
-    }
-    /// The raw value — only for writing the credential into config.
-    pub fn reveal(&self) -> &str {
-        &self.0
-    }
-}
-
-impl Drop for Secret {
-    fn drop(&mut self) {
-        use zeroize::Zeroize as _;
-        self.0.zeroize();
-    }
-}
-
-impl std::fmt::Debug for Secret {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_tuple("Secret").field(&"<redacted>").finish()
-    }
-}
-
-/// The agent-control verbs surfaced by the dashboard, each sent as a
-/// [`WorkspaceInput::Control`]. All five are live: the Agents tab binds `s`
-/// (Stop), `p` (Pause/Resume, toggled by the row's current status), and `r`
-/// (Restart), and Esc sends Stop for the focused agent. The driver honors
-/// Pause/Resume/Restart/Delete on worker lanes — pause parks the worker at
-/// its next step boundary (never mid-tool), restart respawns the lane from
-/// its retained spec, delete stops the worker if one is live and then takes
-/// the lane's row off the deck for good, spec included — and treats them as
-/// no-ops on the lead, whose interrupt is Esc (Stop).
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum AgentControl {
-    Pause,
-    Resume,
-    Stop,
-    Restart,
-    Delete,
-}
-
-impl AgentControl {
-    pub fn label(self) -> &'static str {
-        match self {
-            AgentControl::Pause => "pause",
-            AgentControl::Resume => "resume",
-            AgentControl::Stop => "stop",
-            AgentControl::Restart => "restart",
-            AgentControl::Delete => "delete",
-        }
-    }
 }
 
 #[cfg(test)]
